Move otel config translation to the otel package
