+++ internal/pkg/otel/configtranslate/otelconfig_test.go
@@ -2,7 +2,7 @@
 // or more contributor license agreements. Licensed under the Elastic License 2.0;
 // you may not use this file except in compliance with the Elastic License 2.0.
 
-package coordinator
+package configtranslate
 
 import (
 	"fmt"
